fix: resume recording on foregrounding if necessary
When using the microphone in a UIHostingController, somehow it is possible
for `onDisappear` and `onAppear` to be called again (after the first time!)
when the application is backgrounded and foregrounded again, even when the rest
of the state of the view persists. Since we stop the recorder in the `onDisappear`
to clean up after ourselves, it is possible that we are only backgrounding the app
and therefore need to continue recording when the app is foregrounded again.