fix: implement deactivate for default plugins
These were not implemented as it was previously assumed that `onDisappear` and `onAppear` would not
be called when backgrounding and foregrounding respectively. The plugins are activated and deactivated
in these methods. Since the deactivate method was not implemented for some plugins, callbacks would be
registered multiple times without cleaning up in between. Now, we make sure to properly implement the
deactivate function for the plugins included in the microphone by default.