feat: add onDisappear plugin API
When using the microphone in a UIHostingController, somehow it is possible
for the view's `onDisappear` and `onAppear` functions to be called again (after the first time!)
when the application is backgrounded and foregrounded again, even when the rest
of the state of the view persists.
The newly added `onDisappear` function allows clients to stop the recording when the view disappears.