fix: memory leak of AttendiTranscribePlugin
Since `AttendiMicrophone` is a struct containing reference types like
`Callbacks`, using it in closures can create strong references to those
reference types that create retain cycles and are therefore not garbage collected,
creating memory leaks. Currently, we move all the references to the `mic` in the plugin
outside of any closure definitions, which seems to fix the issue. In terms of dev
experience, it's not really nice, since you have to define all the methods above the
closures before using them.